Used for loop for filterByWord, check if each item has word in it, and if it does, add it to the (filteredArr)ay.
Finished stretch questions.

Used split(" ").length to find the length of the array in "words"
instead of characters.

Used random num gen between 0 and flavorList length to get a random
flavor, and I also removed the flavor from the flavorList so it would
not be selected again and lead to the same flavor being used twice in
the same list.